requirements.txt
gunicorn
flask
numpy
tensorflow-cpu
Pillow
opencv-contrib-python-headless
matplotlib